feat: add auth loading state handling to RootLayout and clean up ContentLayout
